+++ src/ladb_opencutlist/ruby/worker/cutlist/cutlist_labels_worker.rb
@@ -43,7 +43,7 @@ module Ladb::OpenCutList
       return { :errors => [ 'tab.cutlist.error.no_model' ] } unless model
 
       # Retrieve parts
-      parts = @cutlist.get_real_parts(@part_ids)
+      parts = @cutlist.get_parts(@part_ids)
       return { :errors => [ 'tab.cutlist.error.no_part' ] } if parts.empty?
 
       entries = []