+++ CHANGELOG.md
@@ -7,7 +7,9 @@ in 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 * 7.1.0 (WIP)
 
   * Added NL language
+  * Added "normal" font size value to settings
   * Improved PathFormulaWrapper. It now contains instance objects
+  * Improved export with grouped parts activated
   * Fixed rounding errors for metric dimensions in cutting diagrams
   * Fixed DXF line type error